+++ pynamics/constants.py
