+++ pynamics/example/cylinder_buoy.py
@@ -3,34 +3,31 @@ import os
 import numpy as np
 import trimesh
 
-from pyboatsim.dynamics import Gravity, Buoyancy, QuadraticDrag
-import pyboatsim.kinematics.topology as topo
-import pyboatsim.kinematics.joint as joint
-from pyboatsim.boatsim import Sim
-from pyboatsim.visualizer import Visualizer
+from pynamics.dynamics import Gravity, Buoyancy, QuadraticDrag
+import pynamics.kinematics.topology as topo
+import pynamics.kinematics.joint as joint
+from pynamics.boatsim import Sim
+from pynamics.visualizer import Visualizer
 
 if __name__ == "__main__":
 
     body = topo.Body(
-        mass=500,
-        center_of_mass=np.matrix([0, 0, 0]).T,
-        inertia_matrix=np.matrix([
-            # Point mass
-            [100,0,0],
-            [0,100,0],
-            [0,0,100]
-        ])
+        mass_properties_model=trimesh.load(
+                file_obj=os.path.join("models", "common", "cylinder.obj"), 
+                file_type="obj", 
+                force="mesh"),
+        density=0.8*997
     )
 
     water_world = topo.Topology()
     water_world.add_connection("World", "Identity", body.copy(), "Buoy", joint.FreeJoint())
-    water_world.joints["Buoy"].set_configuration(np.matrix([np.pi/8,0,0 ,0, 0, 1]).T)
+    water_world.joints["Buoy"].set_configuration(np.matrix([np.pi/4,0,0 ,1, 1, -4]).T)
     # water_world.joints["Buoy"].set_configuration(np.matrix([0, 0, 2]).T)
     water_world_vis = Visualizer(
         topology=water_world,
         visualization_models={
             (f"Buoy", "Identity"): trimesh.load(
-                file_obj=os.path.join("models", "cup", "cup_boundary_poked.obj"), 
+                file_obj=os.path.join("models", "common", "cylinder.obj"), 
                 file_type="obj", 
                 force="mesh"),
         }
@@ -43,7 +40,7 @@ if __name__ == "__main__":
             "buoyancy": Buoyancy(
                 buoyancy_models={
                     "Buoy": trimesh.load(
-                        file_obj=os.path.join("models", "cup", "cup_boundary_poked.obj"), 
+                        file_obj=os.path.join("models", "common", "cylinder.obj"), 
                         file_type="obj", 
                         force="mesh" 
                     )
@@ -52,11 +49,11 @@ if __name__ == "__main__":
             "drag": QuadraticDrag(
                 drag_models={
                     "Buoy": trimesh.load(
-                        file_obj=os.path.join("models", "cup", "cup_boundary_poked.obj"), 
+                        file_obj=os.path.join("models", "common", "cylinder.obj"), 
                         file_type="obj", 
                         force="mesh" 
                     )
-                }
+                },
             )
         },
     )
