+++ pynamics/math/linalg.py
