+++ pynamics/models/bo-at/bo-at.obj
