+++ pynamics/models/common/Cube.obj
