+++ pynamics/models/cup/bodydrag_model.obj
