+++ pynamics/models/cup/cup.obj
