+++ pynamics/models/cup/cup_boundary.obj
