+++ pynamics/models/cup/cup_boundary_poked.obj
