+++ pynamics/models/cup/water_wheel.obj
