+++ pynamics/models/cup/water_wheel_paddle.obj
