+++ pynamics/models/link/Link0p25m.obj
