+++ pynamics/models/link/Link1m.obj
